+++ example/integration_test/plugin_integration_test.dart
@@ -10,7 +10,7 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 
-import 'package:media_saver/media_saver.dart';
+import 'package:media_saver_plus/media_saver_plus.dart';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();