+++ test/media_saver_method_channel_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:media_saver/media_saver_method_channel.dart';
+import 'package:media_saver_plus/media_saver_method_channel.dart';
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();