+++ test/media_saver_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:media_saver/media_saver.dart';
-import 'package:media_saver/media_saver_platform_interface.dart';
-import 'package:media_saver/media_saver_method_channel.dart';
+import 'package:media_saver_plus/media_saver_plus.dart';
+import 'package:media_saver_plus/media_saver_platform_interface.dart';
+import 'package:media_saver_plus/media_saver_method_channel.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
 class MockMediaSaverPlatform